@@ -2,17 +2,21 @@
   "name": "react-oauth",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:3001",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^12.0.0",
+    "@fortawesome/fontawesome-svg-core": "^6.3.0",
+    "@fortawesome/free-solid-svg-icons": "^6.3.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "axios": "^1.3.4",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-google-login": "^5.2.2",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
+    "classnames": "^2.3.2",
+    "env-cmd": "^10.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-modal": "^3.16.1",
+    "react-router-dom": "^6.8.1",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.58.3",
+    "web-vitals": "^2.1.4",
+    "zustand": "^4.3.6"
   },
   "scripts": {
     "start": "react-scripts start",
